+++ source/platform/os/linux/tivx_queue.c
@@ -77,7 +77,7 @@ typedef struct _tivx_queue_context {
 
 
 vx_status tivxQueueCreate(
-    tivx_queue *queue, uint32_t max_elements, uint32_t *queue_memory,
+    tivx_queue *queue, uint32_t max_elements, uintptr_t *queue_memory,
     uint32_t flags)
 {
     vx_status status = VX_FAILURE;
@@ -198,7 +198,7 @@ vx_status tivxQueueDelete(tivx_queue *queue)
     return (status);
 }
 
-vx_status tivxQueuePut(tivx_queue *queue, uint32_t data, uint32_t timeout)
+vx_status tivxQueuePut(tivx_queue *queue, uintptr_t data, uint32_t timeout)
 {
     vx_status status = VX_FAILURE;
     volatile vx_bool do_break = vx_false_e;
@@ -279,7 +279,7 @@ vx_status tivxQueuePut(tivx_queue *queue, uint32_t data, uint32_t timeout)
     return (status);
 }
 
-vx_status tivxQueueGet(tivx_queue *queue, uint32_t *data, uint32_t timeout)
+vx_status tivxQueueGet(tivx_queue *queue, uintptr_t *data, uint32_t timeout)
 {
     vx_status status = VX_FAILURE;/* init status to error */
     volatile vx_bool do_break = vx_false_e;
@@ -386,7 +386,7 @@ vx_bool tivxQueueIsEmpty(tivx_queue *queue)
     return (is_empty);
 }
 
-vx_status tivxQueuePeek(tivx_queue *queue, uint32_t *data)
+vx_status tivxQueuePeek(tivx_queue *queue, uintptr_t *data)
 {
     vx_status status = VX_FAILURE;/* init status to error */
     tivx_queue_context context = NULL;
